add market orders, nft deposit and withdrawals

+++ src/api/withdrawal/_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, status, HTTPException, Query
-from loguru import logger
 
 from src.service.admin import AdminService, Const
+from src.service.history import HistoryService
 from src.service.number import NumberService
 from src.service.user import UserService
 from src.service.username import UsernameService
@@ -143,30 +143,33 @@ async def withdraw_ton(data: WithdrawTonRequest):
     msg_hash = await WithdrawalService.withdraw_ton(
         data.user_id, data.address, data.tag, data.amount - fee
     )
-    if isinstance(msg_hash, str):
-        await UserService.add_withdrawal(
-            user_id=data.user_id,
-            ton_balance=user.ton_balance,
-            usdt_balance=user.usdt_balance,
-            token='TON',
-            amount=data.amount,
-            address=data.address,
-            tx_hash=msg_hash,
-        )
-    else:
-        logger.error('Not enough balance')
+    if not msg_hash or not isinstance(msg_hash, str):
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail='Some server error occurred',
         )
+
+    await UserService.add_withdrawal(
+        user_id=data.user_id,
+        ton_balance=user.ton_balance,
+        usdt_balance=user.usdt_balance,
+        token='TON',
+        amount=data.amount,
+        address=data.address,
+        tx_hash=msg_hash,
+    )
 
 
 @router.post(
     '/number',
     responses={
         status.HTTP_201_CREATED: {'description': 'Withdraw user token'},
-        status.HTTP_400_BAD_REQUEST: {'description': 'Specified number not found'},
-        status.HTTP_404_NOT_FOUND: {'description': 'Specified user not found'},
+        status.HTTP_400_BAD_REQUEST: {
+            'description': 'Specified user has not enough balance to pay fee'
+        },
+        status.HTTP_404_NOT_FOUND: {
+            'description': 'Specified user or specified number not found'
+        },
         status.HTTP_405_METHOD_NOT_ALLOWED: {
             'description': 'Specified user has not specified number'
         },
@@ -184,7 +187,7 @@ async def withdraw_number(data: WithdrawNumberRequest):
     number = await NumberService.get_number(data.number)
     if not number:
         raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
+            status_code=status.HTTP_404_NOT_FOUND,
             detail=f'Number {data.number} not found',
         )
 
@@ -194,10 +197,38 @@ async def withdraw_number(data: WithdrawNumberRequest):
             detail=f'User {data.user_id} has not number {data.number}',
         )
 
+    fee = await AdminService.get_constant(Const.FEE_WITHDRAWAL_TON)
+    if not fee:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail=f'Some server error occurred',
+        )
+
+    if user.ton_balance < fee:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f'User {data.user_id} has not enough balance to pay fee',
+        )
+
     msg_hash = await WithdrawalService.withdraw_nft(
         data.user_id, data.address, number.address
     )
+    if not msg_hash or not isinstance(msg_hash, str):
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail='Some server error occurred',
+        )
+
     await NumberService.delete_number(data.number)
+    await UserService.update_ton_balance(data.user_id, user.ton_balance - fee)
+
+    await HistoryService.add_nft_withdrawal(
+        user_id=data.user_id,
+        nft_name=number.number,
+        nft_address=number.address,
+        address=data.address,
+        tx_hash=msg_hash,
+    )
 
 
 @router.post(
@@ -233,7 +264,35 @@ async def withdraw_username(data: WithdrawUsernameRequest):
             detail=f'User {data.user_id} has not username {data.username}',
         )
 
+    fee = await AdminService.get_constant(Const.FEE_WITHDRAWAL_TON)
+    if not fee:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail=f'Some server error occurred',
+        )
+
+    if user.ton_balance < fee:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f'User {data.user_id} has not enough balance to pay fee',
+        )
+
     msg_hash = await WithdrawalService.withdraw_nft(
         data.user_id, data.address, username.address
     )
+    if not msg_hash or not isinstance(msg_hash, str):
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail='Some server error occurred',
+        )
+
     await UsernameService.delete_username(data.username)
+    await UserService.update_ton_balance(data.user_id, user.ton_balance - fee)
+
+    await HistoryService.add_nft_withdrawal(
+        user_id=data.user_id,
+        nft_name=username.username,
+        nft_address=username.address,
+        address=data.address,
+        tx_hash=msg_hash,
+    )
